Refactor Functions.java to remove geotools dependency for most vector functions
