+++ flink/src/main/java/org/apache/sedona/flink/expressions/Functions.java
@@ -17,6 +17,7 @@ import org.apache.commons.lang3.tuple.Pair;
 import org.apache.flink.table.annotation.DataTypeHint;
 import org.apache.flink.table.annotation.InputGroup;
 import org.apache.flink.table.functions.ScalarFunction;
+import org.apache.sedona.common.FunctionsGeoTools;
 import org.locationtech.jts.geom.Geometry;
 import org.locationtech.jts.operation.buffer.BufferParameters;
 import org.opengis.referencing.FactoryException;
@@ -334,14 +335,14 @@ public class Functions {
         public Geometry eval(@DataTypeHint(value = "RAW", bridgedTo = org.locationtech.jts.geom.Geometry.class) Object o, @DataTypeHint("String") String sourceCRS, @DataTypeHint("String") String targetCRS)
             throws FactoryException, TransformException {
             Geometry geom = (Geometry) o;
-            return org.apache.sedona.common.Functions.transform(geom, sourceCRS, targetCRS);
+            return FunctionsGeoTools.transform(geom, sourceCRS, targetCRS);
         }
 
         @DataTypeHint(value = "RAW", bridgedTo = org.locationtech.jts.geom.Geometry.class)
         public Geometry eval(@DataTypeHint(value = "RAW", bridgedTo = org.locationtech.jts.geom.Geometry.class) Object o, @DataTypeHint("String") String sourceCRS, @DataTypeHint("String") String targetCRS, @DataTypeHint("Boolean") Boolean lenient)
                 throws FactoryException, TransformException {
             Geometry geom = (Geometry) o;
-            return org.apache.sedona.common.Functions.transform(geom, sourceCRS, targetCRS, lenient);
+            return FunctionsGeoTools.transform(geom, sourceCRS, targetCRS, lenient);
         }
 
     }
